+++ public/js/changePassword.js
@@ -17,6 +17,7 @@ form.addEventListener('submit', async (event) => {
     });
 
     const result = await response.json();
+    console.log(result);
 
     if (result.error) {
       Swal.fire({